+++ src/lib/Components/Buttons/__tests__/NavigationButton-tests.tsx
@@ -5,7 +5,15 @@ import * as renderer from "react-test-renderer"
 
 import NavigationButton from "../NavigationButton"
 
+import { Theme } from "@artsy/palette"
+
 it("renders properly", () => {
-  const button = renderer.create(<NavigationButton title={"I am a navigation button"} href="/some/path" />).toJSON()
+  const button = renderer
+    .create(
+      <Theme>
+        <NavigationButton title={"I am a navigation button"} href="/some/path" />
+      </Theme>
+    )
+    .toJSON()
   expect(button).toMatchSnapshot()
 })