+++ src/lib/Components/Buttons/__tests__/PrimaryBlack-tests.tsx
@@ -5,9 +5,17 @@ import * as renderer from "react-test-renderer"
 
 import PrimaryBlack from "../PrimaryBlack"
 
+import { Theme } from "@artsy/palette"
+
 describe("PrimaryBlack", () => {
   it("renders properly", () => {
-    const button = renderer.create(<PrimaryBlack text={"I am a shiny new  button"} />).toJSON()
+    const button = renderer
+      .create(
+        <Theme>
+          <PrimaryBlack text={"I am a shiny new  button"} />
+        </Theme>
+      )
+      .toJSON()
     expect(button).toMatchSnapshot()
   })
 
@@ -16,7 +24,13 @@ describe("PrimaryBlack", () => {
       fontSize: 22,
     }
 
-    const button = renderer.create(<PrimaryBlack text={"I am a shiny new button"} textStyle={textStyle} />).toJSON()
+    const button = renderer
+      .create(
+        <Theme>
+          <PrimaryBlack text={"I am a shiny new button"} textStyle={textStyle} />
+        </Theme>
+      )
+      .toJSON()
     expect(button).toMatchSnapshot()
   })
 })