+++ src/lib/Components/Consignments/Components/__tests__/TextInput-tests.tsx
@@ -2,12 +2,36 @@ import React from "react"
 import * as renderer from "react-test-renderer"
 import Text from "../TextInput"
 
+import { Theme } from "@artsy/palette"
+
 it("shows an activity indicator when searching ", () => {
-  const component = renderer.create(<Text text={{ value: "My mocked" }} searching={true} />).toJSON()
+  const component = renderer
+    .create(
+      <Theme>
+        <Text
+          text={{
+            value: "My mocked",
+          }}
+          searching={true}
+        />
+      </Theme>
+    )
+    .toJSON()
   expect(component).toMatchSnapshot()
 })
 
 it("does not have an activity when searching ", () => {
-  const component = renderer.create(<Text text={{ value: "My mocked" }} searching={true} />).toJSON()
+  const component = renderer
+    .create(
+      <Theme>
+        <Text
+          text={{
+            value: "My mocked",
+          }}
+          searching={true}
+        />
+      </Theme>
+    )
+    .toJSON()
   expect(component).toMatchSnapshot()
 })