+++ src/lib/Components/Inbox/ActiveBids/__tests__/ActiveBid-tests.tsx
@@ -3,18 +3,38 @@ import * as renderer from "react-test-renderer"
 
 import ActiveBid from "../ActiveBid"
 
+import { Theme } from "@artsy/palette"
+
 it("renders correctly", () => {
-  const tree = renderer.create(<ActiveBid bid={bid()} />).toJSON()
+  const tree = renderer
+    .create(
+      <Theme>
+        <ActiveBid bid={bid()} />
+      </Theme>
+    )
+    .toJSON()
   expect(tree).toMatchSnapshot()
 })
 
 it("looks right for bids in live auctions that haven't started yet", () => {
-  const tree = renderer.create(<ActiveBid bid={bid(true)} />).toJSON()
+  const tree = renderer
+    .create(
+      <Theme>
+        <ActiveBid bid={bid(true)} />
+      </Theme>
+    )
+    .toJSON()
   expect(tree).toMatchSnapshot()
 })
 
 it("looks right for bids in live open auctions", () => {
-  const tree = renderer.create(<ActiveBid bid={bid(true, true)} />).toJSON()
+  const tree = renderer
+    .create(
+      <Theme>
+        <ActiveBid bid={bid(true, true)} />
+      </Theme>
+    )
+    .toJSON()
   expect(tree).toMatchSnapshot()
 })