+++ src/lib/Components/Inbox/Conversations/Preview/__tests__/ShowPreview-tests.tsx
@@ -4,13 +4,23 @@ import * as renderer from "react-test-renderer"
 
 import ShowPreview from "../ShowPreview"
 
+import { Theme } from "@artsy/palette"
+
 it("renders correctly for a regular show", () => {
-  const tree = renderer.create(<ShowPreview show={show as any} />)
+  const tree = renderer.create(
+    <Theme>
+      <ShowPreview show={show as any} />
+    </Theme>
+  )
   expect(tree).toMatchSnapshot()
 })
 
 it("renders correctly for a fair booth", () => {
-  const tree = renderer.create(<ShowPreview show={show as any} />)
+  const tree = renderer.create(
+    <Theme>
+      <ShowPreview show={show as any} />
+    </Theme>
+  )
   expect(tree).toMatchSnapshot()
 })