+++ src/lib/Components/__tests__/Carousel-tests.tsx
@@ -3,10 +3,16 @@ import * as renderer from "react-test-renderer"
 
 import { Carousel } from "../Carousel"
 
+import { Theme } from "@artsy/palette"
+
 describe("Carousel", () => {
   it("renders properly", () => {
     const sources = [{ imageURL: "https://staging.artsy.net", aspectRatio: 1.0 }]
-    const carousel = renderer.create(<Carousel sources={sources} />)
+    const carousel = renderer.create(
+      <Theme>
+        <Carousel sources={sources} />
+      </Theme>
+    )
     expect(carousel).toMatchSnapshot()
   })
 })