+++ src/lib/Scenes/Home/Components/Sales/Components/ZeroState/__tests__/index-tests.tsx
@@ -3,11 +3,17 @@ import "react-native"
 import * as renderer from "react-test-renderer"
 import { ZeroState } from "../index"
 
+import { Theme } from "@artsy/palette"
+
 jest.mock("WebView", () => "WebView")
 jest.mock("../index.html", () => "")
 
 it("looks correct when rendered", () => {
-  const auctions = renderer.create(<ZeroState />)
+  const auctions = renderer.create(
+    <Theme>
+      <ZeroState />
+    </Theme>
+  )
   expect(auctions).toMatchSnapshot()
 })